requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
requests==2.26.0
Pillow==8.3.2
python-dotenv==0.19.0
deep-translator==1.11.4
langdetect==1.0.9